+++ src/tarman/constants.py
@@ -4,13 +4,13 @@ ITEMS_WARNING = 10000
 HELP_STRING = """Browser window key bindings:
   - c              - create archive from selected files
   - e              - extract selected files
-  - h/?/F1         - help window
+  - ?/F1         - help window
   - LEFT/BACKSPACE - go one directory up
   - q              - quit
   - RIGHT/ENTER    - go in to directory or archive
   - SPACE          - select and unselect files
   - UP/DOWN        - move up or down in browser
-  - .              - toggle show/hide hidden files (".*")
+  - h              - toggle show/hide hidden files (".*")
 
 Overlay window key bindings:
   - ENTER          - confirm/ok